Fix IonLongSerializer.Deserialize()
Method was returning IntValue() when it should be returning LongValue()

+++ Amazon.Ion.ObjectMapper/IonPrimitiveSerializers.cs
@@ -65,7 +65,7 @@ namespace Amazon.Ion.ObjectMapper
 
         public long Deserialize(IIonReader reader)
         {
-            return reader.IntValue();
+            return reader.LongValue();
         }
 
         public void Serialize(IIonWriter writer, long item)
